Upgrades PromiseKit to 4.4.4
I'm seeing some issues when trying to compile OBAKit—the build system is complaining about 'underlying module not found' with regard to PromiseKit, so hopefully a rebuild will fix it. Figured I might as well pick up the latest bug fixes in it while I'm at it. Also, switch back to the classic Xcode build system.